Split integration tests into their own package
