+++ tests/const.py
@@ -1,0 +1,9 @@
+TITLE = 'MyTestAPI'
+VERSION = '0.9'
+BASIC_SCHEMA = {
+    'info': {
+        'title': TITLE,
+        'version': VERSION},
+    'swagger': '2.0',
+    'paths': {},
+    'tags': []}
